+++ tests/unit/test_parser.py
@@ -31,3 +31,13 @@ class TestParser(object):
         mode = "gappyout"
         parsed = parser.parse_args([input_path, "-m", mode])
         assert parsed.mode == mode
+
+    def test_plot_trim_report_with_no_value(self, parser):
+        input_path = "my/input/file.fa"
+        parsed = parser.parse_args([input_path, "--plot_trim_report"])
+        assert parsed.plot_trim_report == ""
+
+    def test_plot_trim_report_with_value(self, parser):
+        input_path = "my/input/file.fa"
+        parsed = parser.parse_args([input_path, "--plot_trim_report", "report.html"])
+        assert parsed.plot_trim_report == "report.html"